+++ server/registerRestDomainRoutes.ts
@@ -15,6 +15,7 @@ import { registerRubrosRoutes } from './routes/registerRubrosRoutes'
 import type { RestRouteContext } from './routes/restRouteTypes'
 import { registerZonasEntregaRoutes } from './routes/registerZonasEntregaRoutes'
 import { registerOrdenesEntregaRoutes } from './routes/registerOrdenesEntregaRoutes'
+import { registerCobranzasRoutes } from './routes/registerCobranzasRoutes'
 import { registerEmpresaRoutes } from './routes/registerEmpresaRoutes'
 import { createDomainServices } from './services/createDomainServices'
 
@@ -50,6 +51,7 @@ export function registerRestDomainRoutes(app: Application, prisma: PrismaClient)
   registerFormasPagoRoutes(app, ctx)
   registerFacturasRoutes(app, ctx)
   registerCobrosRoutes(app, ctx)
+  registerCobranzasRoutes(app, ctx)
   registerReportesRoutes(app, ctx)
   registerZonasEntregaRoutes(app, ctx)
   registerOrdenesEntregaRoutes(app, ctx)